Feat/my profile

+++ src/main/kotlin/com/eatsfinder/domain/comment/service/CommentServiceImpl.kt
@@ -5,9 +5,14 @@ import com.eatsfinder.domain.comment.dto.CommentResponse
 import com.eatsfinder.domain.comment.dto.CommentResponse.Companion.from
 import com.eatsfinder.domain.comment.model.Comment
 import com.eatsfinder.domain.comment.repository.CommentRepository
+import com.eatsfinder.domain.like.repository.CommentLikeRepository
 import com.eatsfinder.domain.post.repository.PostRepository
+import com.eatsfinder.domain.user.model.MyActiveType
+import com.eatsfinder.domain.user.model.UserLog
+import com.eatsfinder.domain.user.repository.UserLogRepository
 import com.eatsfinder.domain.user.repository.UserRepository
 import com.eatsfinder.global.exception.ModelNotFoundException
+import com.eatsfinder.global.exception.profile.ImmutableUserException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -15,7 +20,8 @@ import org.springframework.transaction.annotation.Transactional
 class CommentServiceImpl(
     private val userRepository: UserRepository,
     private val postRepository: PostRepository,
-    private val commentRepository: CommentRepository
+    private val commentRepository: CommentRepository,
+    private val userLogRepository: UserLogRepository
 ) : CommentService {
 
     @Transactional(readOnly = true)
@@ -37,11 +43,17 @@ class CommentServiceImpl(
             "post",
             "이 게시물 아이디: (${postId})는 존재하지 않습니다."
         )
-        commentRepository.save(
-            Comment(
-                content = req.content,
+        userLogRepository.save(
+            UserLog(
                 userId = user,
-                postId = post
+                postLikeId = null,
+                commentLikeId = null,
+                commentId = Comment(
+                    content = req.content,
+                    userId = user,
+                    postId = post
+                ),
+                myActiveType = MyActiveType.COMMENT
             )
         )
         return "댓글이 작성되었습니다!"
@@ -53,6 +65,13 @@ class CommentServiceImpl(
             "comment",
             "이 댓글(${commentId})은 존재하지 않습니다."
         )
+        if (comment.userId.id != userId) {
+            throw ImmutableUserException("이 댓글을 삭제할 권한이 없습니다.")
+        }
+
         commentRepository.delete(comment)
+        val userLog = userLogRepository.findUserLogByCommentIdAndMyActiveType(comment, MyActiveType.COMMENT)
+            ?: throw ModelNotFoundException("userLog")
+        userLogRepository.delete(userLog)
     }
 }
